Add support for providing credentials via system keyring
* Move `get_access_token()` to separate `auth` module (imported in `rest_api` module for backwards-compatibility)
* Add functions to set and retrieve creds from keyring
* Add more examples & info to docs
