+++ pyinaturalist/exceptions.py
@@ -1,7 +1,9 @@
-# TODO: Make these extend `requests.HTTPError` to simplify error handling in client code?
-class AuthenticationError(Exception):
+from requests import HTTPError
+
+
+class AuthenticationError(HTTPError):
     pass
 
 
-class ObservationNotFound(Exception):
+class ObservationNotFound(HTTPError):
     pass
