+++ setup.cfg
@@ -20,6 +20,12 @@ classifiers =
 [mypy-forge]
 ignore_missing_imports = True
 
+[mypy-keyring]
+ignore_missing_imports = True
+
+[mypy-keyring.errors]
+ignore_missing_imports = True
+
 [mypy-pytest]
 ignore_missing_imports = True
